requirements.txt
fastapi
uvicorn
pydantic
requests
stability-sdk
Pillow
transformers
torch
sentencepiece
sacremoses
